Purpose of activity is now dynamic
Addressing the feedback from user research, during the creation of an activity
the form step for 'Purpose' needed to be dynamically populated with the level
of the activity. That would work as a reminder for users that what they are
creating is either a fund, a programme or a project.

In order to achieve this, I have added that level to the yaml file so it
could be displayed in the view for the purpose step. After doing that and checking
all the tests were passing and the view on the browser was actually what I wanted
to be, I realised that the page title on the browser tab was displaying
'Purpose of %{level}' instead of being dynamically populated. I believe this page
title comes from a meta tag that it is common to the entire application. I think
changing this meta tag would be too disruptive, so I decided that the best option
would be to create another entry on the yaml file for
'page_title.activity_form.show.purpose_level' where I could specify the level
of the activity and pass that to the view and the tests. This allows me to leave
'page_title.activity_form.show.purpose' as only 'Purpose' so the browser tab
will not display a literal interpolation anymore.

I modified the tests we already had to make sure the user will see the appropiate
information.