Change how organisations are managed
Like managing users, managing organisations is now in the main
application navigation and so is a top level page.

Display the list of organisations as a table with show and edit links,
which makes it consistent with the users view.

Remove the back link on the organisation index view as it is now at the
top level.