Use our own script to the install the cloudfoundry provider
The install script provided by the provider authors relies on the github
API being stable. In practice we have found this not to be the case.
Therefore we are going to use curl to download the version we want and live with having to update this if a new version is released.