Use DateHelper to reject any dates of which one part is 0
If a date contains a part which is only 0 (e.g. 10-0-2020) DateHelper will reject
it before the controller attempts to assign it to an Activity.

We noticed Papertrail recording ActiveRecord::MultiparameterAssignmentErrors when a user input 0 into any of the three inputs on any date field (generated via govuk_date_field)

Rails will attempt to coerce any values from a date_field view helper into a date. If any of those values are 0, the coercion will fail with ActiveRecord::MultiparameterAssignmentErrors

Unfortunately, because this coercion happens before the values are assigned to the model, we cannot catch them with a validator on the model and return errors on the attributes in question.

So we have used DateHelper#format_date (which was previously not being used!) to
weed out any dates which have 0 as one part, and reject them before the controller
attempts to assign the three values as a Date object to the model.

This is was second attempt at this, previously we opened and closed an unmerged
PR https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/pull/258
in which we attempted to add a more specific error message to the field, saying
that it was invalid because of a zero value, not just "blank". However these
attempts were hacky at best and we decided to not go with them.