Update deployment instructions to include the same manual merge step
We tried using a hyrid of these instructions and it went wrong.

The release commit that was merged into master wasn't pushed back downstream into develop. When we came to release again our changelog was in the state as if the first release hadn't happened - not ideal.

We like pull requests because they protect us from accidentally doing git merges to the wrong environment which we use for development. That said, having a pull request to deplo and another to push the change back to the downstream involves a frustrating amount of overhead. Getting 2 pull requests, merged in the right order, with 2 reviews. It's complicated and things are likely to go wrong too.

We're going to keep exactly to the deployment process of RMI instead of using a hyrbid https://crown-commercial-service.github.io/ReportMI-service-manual/#/deployments and go from there.