Update feature specs for organisation page changes
The changes to the organisation show page means these feature specs need
updating to reflect the different steps a user takes through the
application.

In tests, when creating a programme, it must be assigned the extending
organisation of the user under test, otherwise the user will not see
it due to the new programme policy.