Fix the feature tests so that they run with against the active record session store
After the recent change to the way we store session data this highlighted a problem with our session stubbing that we've been vaguely aware of.

By using rack_session_access we were stubbing authenticated session data rather than mocking OAuth. Mocking OAuth is more desirable as it tests the application to the end of the boundary, that is allowing the app to set sessions against realistic OAuth behaviour.

The cost of this change is in increased time that the test suite takes to run. Ever feature test must go to the home page and click the sign in button before continuing. This is more realistic which is important but over enough feature tests these may add up into slower tests which is undesirable.

From discussions with Meyric we think that one way to chip away at the risk of a slow test suite is to move more tests from feature tests into request specs that are faster. Whilst I don't think this is the time to do this refactoring, I do think it's important to note how this change fits into a general direction.