Redis is available to CI when running the test suite
This accurately reflects the state of production.

NB. Without this change Travis seems to already be making an environment called `REDIS_URL` available that refers to `redis://localhost`. This is odd since we haven't declared is in the `- services` list. Instead of worrying too much about where this is coming from, I have been explicit in setting the Redis dependency. This should make it easier to read and understand what Travis is doing.