Use Sidekiq-cron to run the session:trim rake tasks
ActiveRecord session store strongly advises that we trim old sessions that have expired and are no longer used, in order to keep the table size reasonable [1].

This adds a platform for adding other types of scheduled job. As GPaaS doesn't yet have an easy way to add these this is something that we are going to run into soon.

The cron expression is set to run on the first of every month. This seems reasonable given the number of users of the service is expected to be below 50. We could look at making it run less frequently however there is an argument that if we have it running semi-frequently we are likely to notice if it's failing and be able to fix the problems earlier in beta.

[1] https://github.com/rails/activerecord-session_store