Sector step improve content design
trello: https://trello.com/c/GL8qRPuM

Implemented content design suggestions for the sector step in creating an activity.

The page title and heading 1 has changed to be ‘Purpose code’. Also there’s a more descriptive content for the label and hint text.

These suggestions are made after several rounds of testing with users, from which we’ve learnesd that a ‘Sector’ is not descriptive or indicative to the users, and they are more familiar with using ‘Purpose codes’.

As next we’re planning to replace the very long list of select options with a 3 digit code level categories list, and after that to present the user with a more specific sector list after choosing the main category. Before we implement this we need to test this approach first.