Unknown Auth0 error messages are not displayed
This work came out of the Pentest which was performed on the service in
March 2020.

An attacker could inject their own error message into a fake
Auth0 response and get the user to inteact with a bogus service.

By checking the error message is something we know and understand and
then displaying our own error message instead we prevent this attack.

If we see a unknown error message, we log to Rollbar so we can learn
what errors we are seeing and display a generic error message.

As ever, I struggled to find a good home for the locales that display
the errors, I went for a location that already had 'errors' and 'auth0'
over creating more separation.

I reworded some of the specs to better reflect what is under test.