Sector step renamed to Focus Area for clarity and hint text added to the form
The form step changes dynamically to include the level of activity being created. Once this contextual guidance has been used to create the activity, the static "Focus area" is used instead of "Sector" through the rest of the service (apart from IATI where sector is used).

"Focus area in {level}" is only used within the form when that concept is introduced to users. Once introduced the theory is that users can link this back when they see the content within the service and the dynamic inclusion becomes redundant. For example, viewing a fund level activity and seeing "Focus area for your fund" in the list of contents

Co-authored-by: Agz <agz@dxw.com>