Organisation factory only create one BEIS
There should only ever be one BEIS organisation. Our organisation
factory was able to create multiple instances of a BEIS organisation
under different conditions.

This came about after seeing a flaky spec fail and then pass, we
originally fixed this by adding explicit fund setup to the spec, whilst
this worked, it introduced an additional burden on the team to always
build the activity hierarchy correctly, even when elements built are not
under test. We prefer the confidence of knowing the factory will reflect
reality and the focus of the spec on what is under test.

The problems with this are:

- it does not reflect the reality
- because three are created and we cannot guarantee the order we could
  be asserting against the wrong organisation instance
- we get a minor performance improvements but not running multiple
  creates

We had to solve the problem of how to test the logic we were adding to
the factory without relying on "puts Organisation.count". There is no
natural home for Factory tests so we created one at the root of the spec
folder. We had tried to add it into spec/factories/spec/ however this
caused the test to be run twice!

Co-authored-by: Tom Hipkin <tomhipkin@gmail.com>