Anonymize user's IP address when logging to Rollbar
Much of this is lifted from https://github.com/DFE-Digital/teacher-vacancy-service/

We are currently logging users' IP addresses in full to Rollbar, which is a
privacy concern. This PR uses an anonymizing method to strip the last octet from
IP addresses so that they are still recognisable for tracking issues, but cannot
be used to trace the exact location/information of a user.

Note that the result from the test anonymising the v6 IP address lacks the final
`":0000:0000:0000:0000:0000"` because these can be ommitted
See https://en.wikipedia.org/wiki/IP_address#/media/File:Ipv6_address.svg