Country list reduced
When creating an activity, in the recipient country step there was a list including
all the countries on the ISO list.
Following advice from users regarding this list being too extensive and
including countries that ODA do not consider recipients, I have reduced the
list of countries leaving only the ones that ODA report as recipients.