Finance form field and references removed
When a user creates an activity we are now defaulting the finance step to be code
110 "Standard grant", as per user research feedback, since this is the only code
under which activities are reported to IATI.

The user now does not need to introduce a value for this step. It will be assigned
automatically when an activity is created.

I have removed the finance step from the controller (wicked wizard) and deleted the
activity form partial for the finance step to make sure the user is not prompted
to include it. I have also modified the activity show so it `Finance` will not appear.

In a similar way, the references to Finance on the yaml files have been deleted.
However, Finance codes can still be found on a vendor yaml file for reference.