(Fix) Parse ENV["DOMAIN"] to retrieve host name for use in config.hosts
In https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/pull/286
we enabled Rails 6 host whitelisting. Unfortunsately in that PR we used the
ENV["DOMAIN"] variable as the host, but this contains a full URI and not just
the host name. This PR parses the DOMAIN variable to extract only the host for
use in `config.host`