Default value added on Finance
I have modified the value on Finance step as now it is a default value that the
user does not need to enter. The default code for this value is 110, which belongs
to "Standard grant". For this I have created a constant and a method on the activity model
to pass this value every time an activity is created.
This was a request from user research feedback, as this code is the only code
they use when reporting activities to IATI.
I added tests to assert we will always have the right value added to the finance activity.

An entry in the CHANGELOG.md file has been made.