Add JavaScript testing for Capybara
Using Firefox and Geckodriver[1] Capybara can test JavaScript features
in the application.

Configure Capybara to use headless mode so the browser window is not
shown during tests.

Developers will need to install both Firefox and Geckodriver to run the
test suite from this point forward, document this change in the readme.

The Capybara `always_include_port` allows both the rack_test and the
browser version to report the correct port number. As far as I can tell,
when a puma instance is booted up to run the browser tests a random port
number is used and this is the only way to keep this behaviour and still
get the port number right to carry out the tests.

[1] https://github.com/mozilla/geckodriver