Add testing stage to Dockerfile
Add a new test stage to the Dockerfile to install both Firefox and
Geckodriver so our feature specs can test JavaScript in CI.

Install the ESR version of Firefox via apt.

Geckodriver must be installed manually.

The version of Geckodriver install can be set using an argument during
the build stage, it defaults to 0.26.0, the latest at this time.

Also added some comments to show each stage clearly.

As we now have both the release stage and the test stage we can only
push the release stage to Docker Hub for deployment.