Transaction and budget Pundit policies now check whether users can interact with them
Before this change these policies were set to true by default. This is likely due to the fact that transactions and budgets couldn't be viewed so the create and edit buttons were hidden. This wouldn't have stopped fund managers being able to craft URLs to create new transactions on projects so wasn't ideal.

Now that we are showing transactions and budgets to both fund managers and delivery partners we need a way of hiding the create and edit functionality properly.

This change adds the logic into both policies to allow this. The approach is to allow creation and edit of transactions and budgets if the user has the same priviledge on the parent activity. If the user can create a project they should be able to create child attributes such as transactions - the same for updating.

Co-authored-by: Laura Porter <laura@dxw.com>