`Tied status` has now a default value
After discussion with service users, they informed us that the only category they
use for `Tied status` is `Untied` (code "5"), hence there is no need for a form step when
they create activities. This is unlikely to change in the future so we are confident
that removing this step will only be benefitial for the users as they will be able to
complete the journey quicker without affecting the quality of the documentation.

The default value comes from a constant in the activity model now. I added a test to
assert the data will always be included in the XML