`Tied status` form step removed
According to feedback from service users, they only use `Untied` for the `Tied status`
category and this is likely to continue in the future. For this reason, we have agreed
to remove this step completely from the form when a user creates an activity and
make its value a default for `Untied`.

I have deleted the form partial for this step and modify other views and controllers
accordingly. I have also deleted the references to tied_status on yaml files. However,
there is still a file in the vendor folder with the codes for reference.