Document how to backup and restore an environment with production data
The instructions are largely based on the GPaaS documentation for this [1].

We found an issue during this work where the `psql < dump` command had to be run multiple times in order to overcome database constraints. Activities could not be created where extending organisations did not exist. Instead of trying to make any changes to the schema constraints which would require us to remember as we add more, we can ignore the errors by dropping the whole database including the schema and schema_migrations table. By doing this I believe the dump restore command works reliably on a single execution.

Since the data from the dump was originally owned by a role called rdsbroker_<uuid>_manager we need to set that role in the target database so that it is present before the data that links to it is added. Without this the restore command fails with many role errors.

[1] https://docs.cloud.service.gov.uk/deploying_services/postgresql/#paas-to-paas