Ensure that organisation iati_references are unique at a database level
During the production database restore to the pentest environment I noticed that if you skip activerecord you can have as many organisations with GB-GOV-13 as you like. This is a pretty important field to be unique for publishing to IATI purposes to lets validate at the database level as well as ActiveRecord.