Validate that iati references are prefixed with the correct structure
Following the IATI guidance on organisation identifiers: http://reference.iatistandard.org/203/organisation-identifiers/#two-components-of-iati-organisation-identifiers. This change validates that the "namespace" code exists as a prefix and is separated by a hyphen.

Whilst IATI do say that each component of the namespace code MAY also include a hyphen the real world examples we have seen for BEIS do not. It feels like we can layer on this logic later if required.