Add public_activity gem
Because `public_activity` creates a table called Activities by default, and tracks
events using an Activity model, we need to change the name to something else to
avoid a conflict with "our" activities. So we have configured the gem to use
`auditable_events` as its model/table name to avoid confusion.