Use accessible autocomplete to enhance country step
When JavaScript is available, the country select box is progressively
enhanced into a combo box. A user can select from the list of countries
or type the name of a country to fill in the field.

Out of the box the Accessible Autocomplete does not apply a font style
to the options list so we add this locally.

There was an issue with the drop down arrow being rendered under the
body element which is an issue, see:

https://github.com/alphagov/accessible-autocomplete/issues/351

We've used the work-around for now.

The JavaScript is included inline as it is only used for this one
step, it was felt this tightly couples the script and it's use to the
view. If more views use the script, a different approach can be used.

Whilst working on the Recipient country step the content was also
updated to reflect the latest content design: Recipient county becomes
Country.

To test this the `country_select_options` helper was updated to work in
the test suite.

It is worth noting that no validations were updated as part of this
work. When a user has completed all the activity steps and changes the
geography, there is still missing validation on the autocompleted field,
there is separate work to resolve that issue.