Track auditable events on User creation
A first pass at tracking user create actions.

When a user is created, PublicActivity::Activity records are created with the
keys "user.create", "user.update" (because the user is first created in our database,
then an Auth0 identifier is generated, and the user is updated with the Auth0
identifier.

We also rejigged some of the test stubs to be more generic, for re-use in a new
test to verify that the public_activity gem is working as expected.