Tied_status column removed from activities table
Since the user is not going to select a value for `tied_status` when they create activities,
after discussion with other developers, we decided to remove this column from the
DB as it does not serve a purpose anymore. No values will be stored in this column.
From now on Tied status is a default value given by a constant on the activity model.