Update deployment docs
During release 3 it was noted that some of the documented steps would
error.

It was also felt that tagging the commit that merges the release into
develop would be a more accurate reflection of the release code.

The example link also had the release the wrong way around which had the
potential to confuse developers or result in an incorrect link being
pasted to the change log.