CI and local tests use the same env vars
This work came out of seeing the tests fail in CI with seed 10466,
running the same seed locally, the tests passed.

Trying to figure out what was different it became apparent that in CI
the `.env.example` variables are loaded and locally the `.env.test`
variables are loaded.

One of the differences was that REDIS_URL is not defined in .env.test
whereas it is in .env.example.

Adding the REDIS_URL to the .env.test file results in seed 10466 failing
locally, which confirms the issue.

This could be fixed two ways but the decision was made to change CI to
use the .env.test variables and add the REDIS_URL in. This means
developers will need to run redis locally now in order to run the test
suite, but also means CI and the local tests run in the same environment
- this feels like a good trade-off.

With both local and CI test failing the same the actual failing test was
examined:

`spec/requests/users_can_sign_out_spec.rb`

This failure seems to relate to test order as the spec will pass with
other seeds. Seed 10466 runs this spec first.

This looks to be the only spec that stubs the env variables.

Stubbing the entire ENV means that other parts of the code that might
use the ENV will fail, and indeed this was the case.

As the the Code Climate gem is already used, it make sense to use it
here to only stub the env variables needed for the test rather than the
whole thing.