Activity records at the identifier step can be updated
We found the problem on production where activity records existed in the database in a state that needed to be manually updated.

Specifically activities of all steps now required a `geography` field but all records had this value set to `nil`.

I presume this is related to the problem Adam found when testing on staging that some activity records could no longer be updated.

This approach adds a new "blank" step of the activity creation. The intent is that this step exists so that we can disable the validations for the `identifier` field when no value exists.