Use the DataMigrate gem to update the geography field on all activities
As this is not a schema change, nor dependent on a schema change I've chosen to use the data_migrate gem to help us run arbitrary code on live environments in a controlled manner.

We can gain a rails console on each environment and run rake commands using the docs/console-access.md within the repo.