Add geography data for exsiting records that have a region
When we added the `geography` field we didn't set values on our environments. None of our 35 activities on production have a value for geography. https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/pull/249/files

An impact of this is that activities that are at the country step or beyond can no longer be edited.

Using a rake script we update the geography to `recipient_region` for all actiivty records that have a region but not a country.

Whilst no activities currently have a country on production. This is not guaranteed to be the case by the time this task is run, so there is a guard to ignore those records. New activities that set a country should set the geography field correctly in any case.