Empty optional dates are not incuded in XML
This change fixes a bug we detected when validating XML files on IATI public
validator. Before, the XML would include the fields for the optional dates
(`actual_start_date` and `actual_end_date`) even when they were left blank by the
users, causing a validation error on IATI validator. Now, the optional date fields
that are left blank by the users will not be included on the XML and the validator
will not give an error anymore.