Travis now reports changing deploy steps to Slack
Travis will now report when builds and deployments to staging and master change to failed or succeeded. This will allow the team to see where an expected deployment has failed and to take the appropriate action to correct the issue.

Travis notification documentation that was followed can be found here https://docs.travis-ci.com/user/notifications/#configuring-slack-notifications

It has been set up to post to beis-oda-events in the dxw slack channel.

The Travis encryption steps `travis encrypt` were used locally as advised to make the change to the secure token. It seems that this action also updated the provider script condition to wrap it in quotes. I'd like to start by trusting the travis gem to format the yml correctly, one to check.