Run data migrations on start before schema migrations
We found that deploying a batch of code to production that the master branch was much further behind in terms of outstanding schema changes. This meant that a data migration that required a much earlier schema change worked locally and on staging, but failed on production.

As schema changes should include the required and relevant data changes we think we can switch the ordering.