Changes on extending organisation
While working on a different feature, I have encountered an error on the associations
for extending organisations on the factories. For all of the activities (fund, programme
and project) the extending organisation is BEIS. After discussion with members of
the team, we realise that BEIS is extending organisation for funds only.
For programmes and projects, delivery partners should be the extending org.

I ammended the associations for those two activity levels, which resulted on a failing
test where the user tries to edit the extending org and leaves it blank. After my
changes, it would not be blank anymore as the association has been created on the
factory. To ammend this test and to give a more real-life example of a programme
for this test, I decided to assign the extending org to nil at the begining of
the context of the test. I feel this situation reflects reality better, since the
test is trying to assert a user can assign and edit a extending org to an activity.