Activities can have a previous identifier
When an activity has already been published to IATI the identifier must
not change. As we are now standardising the identifier we may have to
use two identifiers for some activities.

As existing activities will be imported, do not expose the previous
identifier attribute to users.

Where an activity has a previous identifier use it as the
`iati-identifier` in the xml and output the new identifier using the
`other-identifier` attribute, see:

http://reference.iatistandard.org/203/activity-standard/iati-activities/iati-activity/other-identifier/