public_activity update events on Activities are more granular
Users update an Activity via a multi-step form, a step for each attribute. The
previous implementation of public_activity meant we had an `activity.update`
event for each step. This change changes the event keys to be more obvious
*what* is being created or updated.

We decide if an activity attribute is being created or updated by the `wizard_complete?`
method on Activities - if the wizard is complete, we can assume an attribute is being
updated; if not, we can assume the attribute is being created.

So the event keys will now be `activity.action.attribute`, e.g. activity.create
for the initial Activity creation; activity.create.identifier for the identifier
creaton, and activity.update.identifier for when the identifier is updated.