Do not track the owner of auditable events in the model
Previously we were tracking the owner of auditable_events in the model, as suggested
by the gem's authors. However this involved referencing the controller in the
model, which is not great practice. It also meant that running db:seed failed,
because the seeds try to create model records without a controller in play.

We are now tracking auditable events in the controllers, which means we can more
easily pass the `current_user` as the event owner.