Seeds file creates valid activities
While working on a different feature, I noticed our seeds file does not create
activities that pass the IATI validator, as they were missing important bits of
information in regards to extending organisations and reference numbers.

This commit introduces some changes to the activity seeds file, where now all the
activities created will pass the IATI public validator. In addition to a fund,
I have created two programmes that belong to the fund, and one project that belongs
to one of the programme. This should help us to test the app faster, since we do
not have to create programmes manually every time.