Sector question uses radio buttons
User research showed the sector question was difficult to answer due to
the large number of possible responses. We tested breaking this question
into two steps which made the task much simpler for users.

This is the first step towards breaking the question into two parts.

First the codelist helper is able to return only the sectors in a given
sector category, when no category is supplied, return all sectors.

Then the sector question is presented as radio buttons.