Patch public_activity initializer to set table_name
We followed the author's instructions to change the name of the table public_activity
uses https://github.com/chaps-io/public_activity/wiki/%5BHow-to%5D-Change-table-name-for-activities

However while our test suite passed, when trying to use the application we found that the
table name was still `activities` when we intended to use `auditable_events` (see
d1d4e47efa5c5f1f94572ea7702e938b3b015c05 in this branch)

After much digging we concluded that the way the table name is set in public_activity's
config is not thread-safe - despite it being correctly set on application load,
the table name was later overwritten with `nil`, which then defaulted to `activities`
and which meant our application did not run correctly.

public_activity is only thread-safe if you're using the default table name of `activities`,
not any other table name.

We located the issue at public_activity/orm/active_record/activity.rb:24

In lieu of submitting a patch to public_activity, because of time pressures we have patched
the config in our own initializer. This ensures the default name for the table in
our application is `auditable_events` and it *never* falls back to `activities`