BEIS users only see an organisation's activities on that organisation's show page
BEIS users have the right to view all activities, no matter which organisation they belong to.
The can also view any organisation's show page.

However, because of a bug, BEIS user's would see *all* activities on any organisation's
show page, because the list of activities is not filtered by organisation.

Now, when a BEIS user views an organisation's show page, they will only see activities
which belong to that organisation. That is, fund and projects which belong to the organisation,
and programmes which have the organisation as its extending organisation.

When a BEIS user views the BEIS organisation, they will see *all* activities belonging
to all organisations.

Non-BEIS users will see the same activities on their organisation show page as they
did before.