h1 heading error fixed on Geography step
We found a typo on the heading of the geography step of activities. It should be
a `h1` but it was `hq` instead. As a result, screen readers were struggling to
determine the layout of the page and the accessibility of the app was compromised.

This commit fixes that typo on the heading and hopefully now the page is fully
accessible for people with vision impairment.