Message on links that open in new window
The Accesibility Audit report stated that links that open in new windows should
have a message informing the user of this action. Visual impaired users can find
confusing when a new window opens without informing them (they would not be able
to click on the `Back` button, for example).

In line with the suggestion given by the Accessibility report, I have amended
the messages on the links that open in new windows. Now, all these messages
inform the user they will be have a new window/tab opened as a result when
clicking the link.

I have amended a few test on the `form_helpers` to assert the new message is
correctly displayed on the app.