Parent activities are included in a project XML
IATI contemplates the possibility of publishing the level of hierarchy by
including the parent activities linked to an activity. This commit modifies the
XML template for activities to include a new field called `related-activity` which
will contain the identifiers of the parent activities of a project (fund and
programme). For now, we are only including parent activities for projects as
they are the only ones that are published to IATI at the moment.

I have included tests to assert the new field will always contain the right data.