Add Skylight service for application monitoring
Add the Skylight gem & service to monitor performance.

The SKYLIGHT_ACCESS_TOKEN has been added to Travis CI and is in the 1Password
vault for reference.

The SKYLIGHT_ENV variable has been set to `staging` for the develop branch on
Travis CI, and `production` for the master branch.

Skylight data will be logged under the `accounts-skylight@dxw.com` account on
https://www.skylight.io/