Refactor activity validation to use contexts
Using built in Rails validation contexts makes the interface for the
form steps much simpler.

https://guides.rubyonrails.org/active_record_validations.html#on

Overriding the Activity `valid?` method lets us run all the validations
when certain conditions are met. For now this will be the
`wizard_status` is complete, which means all of the form steps have been
completed and the context is nil, which means any call to `valid?`
without providing a context.

This commit makes the test suite go red, but it was felt better to keep
the commits focused and conscience fixing the sepc in the next commit.

Co-Authored-By: Dominic Baggott <dominic.baggott@gmail.com>