Update geography question
Now that the interface for validation is simpler the dependent geography
steps start working as expected: Users choose to change the geography
answer and then choose the region or country based on their response.

In order for these steps to work we must reset the dependent answers to
nil when the user saves the updated geography answer and update the
specs to reflect the new journey.