Validate activities on create
Previously we skipped validation in order to create an activity via the
form steps in the application.

Now the validations are fired at the appropriate times (during form
steps and once the form is completed) we can validate the activity upon
create.

Activities can be created with an `identifier` of nil through the form
steps so we allow duplicate values of nil, but any other value must be
unique.