Only run data migrations in production
We use the data migrate gem to manage data migrations.

https://github.com/ilyakatz/data-migrate

Previously we ran the data migrations in all environments which caused
an issue when the name of a column in the database was later changed.

Because the test environment destroys and re-creates the database from
the schema, not the migrations, when the data migrations run the
database has a different column name and so fails.

We shouldn't run data migrations in the test environment anyway as there
is no data to migrate.