Refactor IATI fixture so that it only includes a single activity
This should cause no tangible change but for tests that use it the speed at which they can run should be improved. Creating a single activity is much faster than creating multiple, when those extra activities are not the subject of the test there is no value in ingesting them.