Debugging to check why userinfo isn't always set in the session
Only replicatable on staging so we have to push to develop.

On our initial debugging 048c0fd2ffde499a483ed78f7d08ac1b4f6afa43 we learnt that the user does have a session, keys are present but not the userinfo key. Let's verify if the callback endpoint recieves this data when auth0 sends a callback.